Move demo code files into sample project
