+++ BridgeAppSDKSample/ActivityViewController.swift
