+++ BridgeAppSDKSample/AppDelegate.swift
@@ -1,15 +1,100 @@
-//
-//  AppDelegate.swift
-//  BridgeAppSDK
-//
-//  Created by Erin Mounts on 1/28/16.
-//  Copyright © 2016 Apple, Inc. All rights reserved.
-//
+/*
+Copyright (c) 2015, Apple Inc. All rights reserved.
+
+Redistribution and use in source and binary forms, with or without modification,
+are permitted provided that the following conditions are met:
+
+1.  Redistributions of source code must retain the above copyright notice, this
+list of conditions and the following disclaimer.
+
+2.  Redistributions in binary form must reproduce the above copyright notice,
+this list of conditions and the following disclaimer in the documentation and/or
+other materials provided with the distribution.
+
+3.  Neither the name of the copyright holder(s) nor the names of any contributors
+may be used to endorse or promote products derived from this software without
+specific prior written permission. No license is granted to the trademarks of
+the copyright holders even if such marks are included in this software.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE
+FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+*/
 
 import UIKit
-import BridgeAppSDK
+import ResearchKit
 
 @UIApplicationMain
+public class AppDelegate: UIResponder, UIApplicationDelegate {
+    
+    let healthStore = HKHealthStore()
+    
+    public var window: UIWindow?
+    
+    var containerViewController: ResearchContainerViewController? {
+        return window?.rootViewController as? ResearchContainerViewController
+    }
+    
+    public func application(application: UIApplication, willFinishLaunchingWithOptions launchOptions: [NSObject : AnyObject]?) -> Bool {
+        let standardDefaults = NSUserDefaults.standardUserDefaults()
+        if standardDefaults.objectForKey("ORKSampleFirstRun") == nil {
+            ORKPasscodeViewController.removePasscodeFromKeychain()
+            standardDefaults.setValue("ORKSampleFirstRun", forKey: "ORKSampleFirstRun")
+        }
+        
+        // Appearance customization
+        let pageControlAppearance = UIPageControl.appearance()
+        pageControlAppearance.pageIndicatorTintColor = UIColor.lightGrayColor()
+        pageControlAppearance.currentPageIndicatorTintColor = UIColor.blackColor()
+        
+        // Dependency injection.
+        containerViewController?.injectHealthStore(healthStore)
+        return true
+    }
+    
+    public func application(application: UIApplication, didFinishLaunchingWithOptions launchOptions: [NSObject : AnyObject]?) -> Bool {
+        lockApp()
+        return true
+    }
+    
+    public func applicationWillResignActive(application: UIApplication) {
+        if ORKPasscodeViewController.isPasscodeStoredInKeychain() {
+            // Hide content so it doesn't appear in the app switcher.
+            containerViewController?.contentHidden = true
+        }
+    }
+    
+    public func applicationWillEnterForeground(application: UIApplication) {
+        lockApp()
+    }
+    
+    func lockApp() {
+        /*
+        Only lock the app if there is a stored passcode and a passcode
+        controller isn't already being shown.
+        */
+        guard ORKPasscodeViewController.isPasscodeStoredInKeychain() && !(containerViewController?.presentedViewController is ORKPasscodeViewController) else { return }
+        
+        window?.makeKeyAndVisible()
+        
+        let passcodeViewController = ORKPasscodeViewController.passcodeAuthenticationViewControllerWithText("Welcome back to ResearchKit Sample App", delegate: self) as! ORKPasscodeViewController
+        containerViewController?.presentViewController(passcodeViewController, animated: false, completion: nil)
+    }
+}
 
-class AppDelegate: BridgeAppSDK.AppDelegate {
+extension AppDelegate: ORKPasscodeDelegate {
+    public func passcodeViewControllerDidFinishWithSuccess(viewController: UIViewController) {
+        containerViewController?.contentHidden = false
+        viewController.dismissViewControllerAnimated(true, completion: nil)
+    }
+    
+    public func passcodeViewControllerDidFailAuthentication(viewController: UIViewController) {
+    }
 }
