+++ BridgeAppSDKSample/ConsentDocument.swift
