+++ BridgeAppSDKSample/DiscreteGraphDataSource.swift
