+++ BridgeAppSDKSample/HealthClientType.swift
