+++ BridgeAppSDKSample/HealthDataStepViewController.swift
