+++ BridgeAppSDKSample/LineGraphDataSource.swift
