+++ BridgeAppSDKSample/ProfileHeaderView.swift
