+++ BridgeAppSDKSample/ProfileStaticTableViewCell.swift
