+++ BridgeAppSDKSample/ResearchContainerSegue.swift
