+++ BridgeAppSDKSample/ResearchContainerViewController.swift
