+++ BridgeAppSDKSample/StudyTasks.swift
